requirements.txt
circle-fit==0.1.3
lsq-ellipse==2.2.1
matplotlib==3.3.3
more-itertools==8.13.0
networkx==2.3
numpy==1.21.5
pandas==1.3.5
scikit-learn==0.21.2
scipy==1.3.0
tulip-python==5.4.0
scikit-posthocs==0.7.0
sklearn==0.0
mlrose==1.3.0